chore: refactor AccountId
